add DOBPermitIssueNow + tests

+++ datasets/models/HPDBuildingRecord.py
@@ -25,7 +25,7 @@ class HPDBuildingRecord(BaseDatasetModel, models.Model):
     streetname = models.TextField(blank=True, null=True)
     zip = models.TextField(blank=True, null=True)
     block = models.SmallIntegerField(blank=True, null=True)
-    lot = models.SmallIntegerField(blank=True, null=True)
+    lot = models.IntegerField(blank=True, null=True)
     communityboard = models.IntegerField(blank=True, null=True)
     censustract = models.TextField(blank=True, null=True)
     managementprogram = models.TextField(blank=True, null=True)
